qml/QMLVideoPlayer: Fix frequent crash related to object desctruction
We probably need to stop the player before we delete the QImage surface object,
otherwise there might happen some racy condition in unlock() method which then
references QImage surface object which is being(or is already) deleted and
testing application crashes on exit. With this fix it's seems a lot stable, at
least it doesn't crash on application exit.

Signed-off-by: Petr Stetiar <ynezz@true.cz>